TST: add testcases for some decorated plot methods
These tests need to be either included in the original test if
they use the same baseline image or use a differetn baseline
image. Using two different test functions and the same baseline
image results in races which kill ghostscript and others when
they happen to access the same image.